+++ deploy-scripts/src/lib.rs
@@ -3,4 +3,5 @@ pub mod constants;
 pub mod deploy;
 pub mod pause;
 pub mod sanity_check;
+pub mod transfer_ownership;
 pub mod utils;
